% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{align}
\alias{align}
\title{align}
\usage{
align(x1, x2, score.matrix, gap.score = -1, type = "local", debug = TRUE)
}
\arguments{
\item{x1}{sequence (a character vector).}

\item{x2}{sequence (a character vector).}

\item{score.matrix}{substitution matrix.}

\item{gap.score}{gap score.}

\item{type}{type of alignment (global, local), default to local.}
}
\description{
aligns two sequences x1 and x2.
}
\examples{
align("AHSSFOO", "GOODFDKD", type = "global")$alignment
}

