% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\docType{methods}
\name{conservationMatrix}
\alias{conservationMatrix}
\alias{conservationMatrix,AAMultipleAlignment-method}
\alias{conservationMatrix,matrix-method}
\title{Compute the conservation score matrix from a MSA}
\usage{
conservationMatrix(object)

\S4method{conservationMatrix}{matrix}(object)

\S4method{conservationMatrix}{AAMultipleAlignment}(object)
}
\arguments{
\item{object}{AAMultipleAlignment alignment object (or a matrix).}
}
\description{
Compute the conservation score matrix from a multiple sequence alignment (MSA) 
according to the method described in the software Jalview (htt://www.jalview.org).
}
\examples{
NULL
}

