% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{conservationMatrix}
\alias{conservationMatrix}
\title{conservationMatrix}
\usage{
conservationMatrix(object)
}
\arguments{
\item{object}{sequence alignment object.}
}
\description{
computes the conservation score matrix from a MSA
}

