% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dcor}
\alias{dcor}
\title{dcor}
\usage{
dcor(x, use = "pairwise", method = "pearson")
}
\arguments{
\item{x}{matrix with numeric values.}

\item{use}{argument passed down to cor().}

\item{method}{method use for computing correlation (default: pearson).}
}
\value{
a dist object.
}
\description{
Compute Pearson correlation distance on the rows.
}
\examples{
NULL
}
