% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{eud}
\alias{eud}
\title{Euclidean distance}
\usage{
eud(x, y)
}
\arguments{
\item{x}{numeric matrix or vector.}

\item{y}{numeric matrix or vector.}
}
\value{
Numeric matrix or vector.
}
\description{
Compute euclidean distance of two matrices or vectors.
}

