% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{getMotifArchSimilarity}
\alias{getMotifArchSimilarity}
\title{getMotifArchSimilarity}
\usage{
getMotifArchSimilarity(object)
}
\arguments{
\item{object}{MotifSearchResult object.}
}
\value{
A matrix with Jaccard index similarity.
}
\description{
Computes similarity between sequence motif architectures.
}

