% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{getMotifArchString}
\alias{getMotifArchString}
\title{getMotifArchString}
\usage{
getMotifArchString(object, return.unique = FALSE)
}
\arguments{
\item{object}{a MotifSearchResult object.}

\item{return.unique}{logical; whether to return unique architectures.}
}
\value{
A list of sequences.
}
\description{
Obtain  a vector of motif architectures per sequence, as a string. Optionally
encode the motif numbers using letter, LETTER and 0:9 symbols (if the number
of motifs is lower than the number of symbols).
}

