% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\docType{methods}
\name{getMotifCoverage}
\alias{getMotifCoverage}
\alias{getMotifCoverage,MotifSearchResult-method}
\title{getMotifCoverage}
\usage{
getMotifCoverage(object)

\S4method{getMotifCoverage}{MotifSearchResult}(object)
}
\arguments{
\item{object}{a MotifSearchResult object.}
}
\value{
A numeric vector with the percentage of sequence covered by motifs.
}
\description{
Obtain the percentage of each sequence that is covered by the motifs matching
to it. This enable us to know how much the motifs are representative of the
features defining each sequence.
}
\examples{
NULL
}
