% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\docType{methods}
\name{getMotifMatrix}
\alias{getMotifMatrix}
\alias{getMotifMatrix,MotifSearchResult-method}
\title{getMotifMatrix}
\usage{
getMotifMatrix(object, simplify = TRUE, filter = FALSE)

\S4method{getMotifMatrix}{MotifSearchResult}(object, simplify = TRUE,
  filter = FALSE)
}
\arguments{
\item{object}{a MotifSearchResult object.}

\item{simplify}{whether to return a matrix of 0 (motif not present) and 1 (motif present).}

\item{filter}{whether remove columns/rows with all zero entries.}
}
\value{
a matrix.
}
\description{
Obtain a congintency matrix with the number of motifs per sequence. Alterna-
tivelly return only presence/absence of motif, and/or only motifs/sequences
with at least one non-zero entry.
}
\examples{
NULL
}

