% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{getMotifsBySeq}
\alias{getMotifsBySeq}
\title{getMotifsBySeq}
\usage{
getMotifsBySeq(object, unique = FALSE)
}
\arguments{
\item{object}{a MotifSearchResult object.}

\item{unique}{logical; whether to return unique motifs per sequence.}
}
\value{
A list of sequences.
}
\description{
Obtain list with a vector of motifs per sequence.
}
