% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{getMotifSimilarity}
\alias{getMotifSimilarity}
\title{getMotifSimilarity}
\usage{
getMotifSimilarity(object)
}
\arguments{
\item{object}{MotifSearchResult object.}
}
\description{
Computes Jaccard similarity between sequence motifs.
}

