% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\docType{methods}
\name{getMotifTotalCoverage}
\alias{getMotifTotalCoverage}
\alias{getMotifTotalCoverage,MotifSearchResult-method}
\title{getMotifTotalCoverage}
\usage{
getMotifTotalCoverage(object)

\S4method{getMotifTotalCoverage}{MotifSearchResult}(object)
}
\arguments{
\item{object}{a MotifSearchResult object.}
}
\value{
A numeric scalar with the percentage of sequence covered by motifs.
}
\description{
Obtain the total percentage of all sequences that is covered by the motifs matching
to it.
}
\examples{
NULL
}
