% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{motifAlign}
\alias{motifAlign}
\title{Align two motifs represented as PWMs}
\usage{
motifAlign(x1, x2, align.method = "local", score.method = "PCC",
  debug = FALSE)
}
\arguments{
\item{x1}{motif 1.}

\item{x2}{motif 2.}

\item{align.method}{align method, one of "global" or "local" (default = local).}

\item{score.method}{scoring method, one of "PCC" or "euclid" (default = PCC).}

\item{debug}{logical; whether to show debugging information.}
}
\description{
Align two motifs represented as PWMs
}

