% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{MotifCompareResult-class}
\alias{MotifCompareResult-class}
\title{MotifCompareResult}
\description{
MotifCompareResult
}
\section{Slots}{

\describe{
\item{\code{info}}{list with result information.}

\item{\code{query}}{AnnotatedDataFrame with sequence information.}

\item{\code{target}}{AnnotatedDataFrame with motif information.}

\item{\code{probabilities}}{list of motif probabilities.}

\item{\code{matches}}{list of motif scores.}
}}

