% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{MotifSearchResult-class}
\alias{MotifSearchResult-class}
\alias{[,MotifSearchResult-method}
\alias{motifNames}
\alias{motifNames,MotifSearchResult-method}
\alias{nmotif}
\alias{nmotif,MotifSearchResult-method}
\alias{nseq}
\alias{nseq,MotifSearchResult-method}
\alias{sequenceNames}
\alias{sequenceNames,MotifSearchResult-method}
\alias{show,MotifSearchResult-method}
\title{MotifSearchResult}
\usage{
\S4method{show}{MotifSearchResult}(object)

\S4method{[}{MotifSearchResult}(x, i, j, ..., drop = FALSE)

nmotif(object)

nseq(object)

motifNames(object)

sequenceNames(object)
}
\arguments{
\item{object}{MotifSearchResult object.}

\item{x}{MotifSearchResult object.}

\item{i}{row number.}

\item{j}{column number.}

\item{...}{further arguments passed down to method.}

\item{drop}{logical; whether to coerce to vector when number of columns equal one.}
}
\description{
MotifSearchResult
}
\section{Slots}{

\describe{
\item{\code{info}}{list.}

\item{\code{sequences}}{AnnotatedDataFrame.}

\item{\code{motifs}}{AnnotatedDataFrame.}

\item{\code{models}}{list.}

\item{\code{ranges}}{RangedData.}
}}

