\name{motifTools-heatPlot}
\alias{motifTools-package}
\alias{motifTools}
\docType{package}
\title{
motifTools: tools to analyze sequence motifs
}
\description{
This packages contains several functions useful to analyze and interpret results from sequence motif tools. Currently this package is focused on the MEME suite.
}
\details{
\tabular{ll}{
Package: \tab motifTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-02-02\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Diego Diez

Maintainer: Diego Diez <diego10ruiz@gmail.com>
}
\keyword{ package }