% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{plotConservation}
\alias{plotConservation}
\title{plotConservation}
\usage{
plotConservation(x, seq.names = FALSE, cluster.row = TRUE)
}
\arguments{
\item{x}{a matrix as obtained with conservationMatrix()}

\item{seq.names}{logical whether to plot seq names}

\item{cluster.row}{logical whether to cluster rows}
}

