% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\docType{methods}
\name{plotConservationMatrix}
\alias{plotConservationMatrix}
\alias{plotConservationMatrix,matrix-method}
\alias{plotConservationMatrix,AAMultipleAlignment-method}
\title{Plots a heatmap with the conservation score for each residue}
\usage{
plotConservationMatrix(x, ...)

\S4method{plotConservationMatrix}{matrix}(x, tree, raster = FALSE,
  tile.color = "transparent", show.tips = FALSE,
  show.consensus = FALSE, size = 8)

\S4method{plotConservationMatrix}{AAMultipleAlignment}(x, tree,
  raster = FALSE, tile.color = "transparent", show.tips = FALSE,
  show.consensus = FALSE, size = 8)
}
\arguments{
\item{x}{a matrix as obtained with conservationMatrix() or an AAMultipleAlignment object.}

\item{...}{arguments passed down to matrix-method.}

\item{tree}{a tree of class dendrogram or that can be coerced into.}

\item{raster}{logical; whether to use geom_raster (best for large alignments).}

\item{tile.color}{color for the tile borders (default: transparent).}

\item{show.tips}{logical; whether to show sequence ids.}

\item{show.consensus}{logical; whether to show consensus sequence.}

\item{size}{text size for consensus sequence.}
}
\description{
Plots a heatmap with the conservation score for each residue colored according
to the scheme shown in Jalview (htt://www.jalview.org).
}
\details{
Sometimes computing the conservation matrix with conservationMatrix
can be time consuming. In that case it might be better to store the output of
conservationMatrix in a variable and pass that to plotConservationMatrix.
}
\examples{
NULL
}
