% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotMatrix}
\alias{plotMatrix}
\title{plotMatrix}
\usage{
plotMatrix(object, tree, fill, color = "transparent", high, high.col,
  bar.percentage = TRUE, plot = TRUE)
}
\arguments{
\item{object}{a matrix object or list of matrix objects.}

\item{tree}{a tree object.}

\item{fill}{color used for tiles.}

\item{color}{color for tile borders.}

\item{high}{vector (factor or character) indicating row groups.}

\item{high.col}{colors usef for each group in high.}

\item{bar.percentage}{logical; whether to show percentages in bar plot (default: TRUE).}

\item{plot}{logical; whether to draw the plot (default: TRUE).}
}
\value{
Returns a grob object invisibly.
}
\description{
plotMatrix
}
\examples{
NULL
}

