% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{plotMotifArchSimilarity}
\alias{plotMotifArchSimilarity}
\title{plotMotifArchSimilarity}
\usage{
plotMotifArchSimilarity(object)
}
\arguments{
\item{object}{MotifSearchResult object.}
}
\description{
Plots heatmap with similarity between sequence motif architectures.
}

