% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomtom.R
\name{plotMotifMatches}
\alias{plotMotifMatches}
\title{plotMotifMatches}
\usage{
plotMotifMatches(x, fill = c("p_value", "e_value", "q_value"))
}
\arguments{
\item{x}{a MotifCompareResult object.}

\item{fill}{the statistic to plot. One of p_value, e_value, q_value (default: p_value).}
}
\description{
Plot a matrix with rows and columns representing motifs in the query and target databases, and the fill
color representing one of the three statistics (p_value, e_value, q_value) measuring the significance of
the similarity between the motifs.
}
\examples{
NULL
}

