% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plotMotifMatrix}
\alias{plotMotifMatrix}
\alias{plotMotifMatrix,MotifSearchResult-method}
\alias{plotMotifMatrix,list-method}
\alias{plotMotifMatrix,matrix-method}
\title{plotMotifMatrix}
\usage{
plotMotifMatrix(object, ...)

\S4method{plotMotifMatrix}{list}(object, tree, fill, color = "transparent",
  high, high.col, bar.percentage = TRUE, plot = TRUE)

\S4method{plotMotifMatrix}{MotifSearchResult}(object, ...)

\S4method{plotMotifMatrix}{matrix}(object, ...)
}
\arguments{
\item{object}{a matrix object or list of matrix objects.}

\item{...}{arguments passed down to methods.}

\item{tree}{a tree object.}

\item{fill}{color used for tiles.}

\item{color}{color for tile borders.}

\item{high}{vector (factor or character) indicating row groups.}

\item{high.col}{colors usef for each group in high.}

\item{bar.percentage}{logical; whether to show percentages in bar plot (default: TRUE).}

\item{plot}{logical; whether to draw the plot (default: TRUE).}
}
\value{
Returns a grob object invisibly.
}
\description{
Plot a heatmap with motif counts per sequence and associated tree and 
barplots with motif-wise percentages.
}
\examples{
NULL
}

