% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMotifMatrix}
\alias{plotMotifMatrix}
\title{plotMotifMatrix}
\usage{
plotMotifMatrix(object, tree, annot, annot.col)
}
\arguments{
\item{object}{a MotifResult object or list of MotifResult objects.}

\item{tree}{a tree}

\item{annot}{further annotations}

\item{annot.col}{colors for the annotations}
}
\description{
Plots a tree and associated heatmap and barplot of the distribution of motifs in the sequences.
}
\examples{
NULL
}

