% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo.R
\name{readFIMO}
\alias{readFIMO}
\title{Read results from FIMO motif search stored in XML format.}
\usage{
readFIMO(file, description = NULL)
}
\arguments{
\item{file}{file XML output file from FIMO.}

\item{description}{character string describing the dataset.}
}
\value{
MotifSearchResult object.
}
\description{
Read results from FIMO motif search stored in XML format and returns a
MotifSearchResult object.
}
\examples{
NULL
}
