% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mast.R
\name{readMAST}
\alias{readMAST}
\title{Read results from MAST motif search stored in XML format.}
\usage{
readMAST(file, description = NULL)
}
\arguments{
\item{file}{XML output file from MAST}

\item{description}{character string describing the dataset.}
}
\value{
MotifSearchResult object.
}
\description{
Read results from MAST motif search stored in XML format and returns a
MotifSearchResult object.
}
\examples{
NULL
}

