% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme.R
\name{readMEME}
\alias{readMEME}
\title{Read results from MEME motif search stored in XML format.}
\usage{
readMEME(file, description = NULL)
}
\arguments{
\item{file}{XML output file from MEME.}

\item{description}{character string describing the dataset.}
}
\value{
MotifSearchResult object.
}
\description{
Read results from MEME motif search stored in XML format and returns a
MotifSearchResult object.
}
\examples{
NULL
}

