% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{readMEME_old}
\alias{readMEME_old}
\title{Read MEME XML file format}
\usage{
readMEME_old(filename, sequenceData, description = NULL)
}
\arguments{
\item{filename}{MEME XML file.}

\item{sequenceData}{sequence data object (optional).}

\item{description}{dataset description.}
}
\value{
MotifSearchResult object.
}
\description{
Read MEME XML file format
}
\examples{
NULL
}

