% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomtom.R
\name{readTOMTOM}
\alias{readTOMTOM}
\title{readTOMTOM}
\usage{
readTOMTOM(file, do.cut = FALSE, cut.col = "q-value", cutoff = 0.05,
  filter = "TRANSFAC", do.clean = TRUE)
}
\arguments{
\item{file}{}

\item{do.cut}{}

\item{cut.col}{}

\item{cutoff}{}

\item{filter}{}

\item{do.clean}{}
}
\value{
MotifCompareResult object.
}
\description{
readTOMTOM
}
\examples{
NULL
}

