% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{align}
\alias{align}
\title{Align two sequences}
\usage{
align(x1, x2, score.matrix, gap.score = -1, type = "local", debug = FALSE)
}
\arguments{
\item{x1}{sequence (a character vector).}

\item{x2}{sequence (a character vector).}

\item{score.matrix}{substitution matrix.}

\item{gap.score}{gap score.}

\item{type}{type of alignment: global or local (default: local).}

\item{debug}{logical; whether to return debugging data.}
}
\value{
AlignResult object.
}
\description{
Align two sequences using local (Smith-Waterman) or global (Needleman–Wunsch) sequence alignment.
}
\examples{
library(Biostrings)
data(BLOSUM62)
align("ALVDE", "AVRES", score.matrix = BLOSUM62)
}

