% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{align}
\alias{align}
\title{Align two sequences}
\usage{
align(x1, x2, score.matrix = NULL, gap.score = -1, type = "local",
  debug = FALSE)
}
\arguments{
\item{x1}{pattern sequence: a character vector, scalar or AAString.}

\item{x2}{subject sequence: a character vector, scalar or AAString.}

\item{score.matrix}{substitution matrix.}

\item{gap.score}{gap score.}

\item{type}{type of alignment: global or local (default: local).}

\item{debug}{logical; whether to return debugging data.}
}
\value{
AlignResult object.
}
\description{
Align two sequences using local (Smith-Waterman) or global (Needleman-Wunsch) sequence alignment.
}
\examples{
library(motiftools)
data(BLOSUM62, package = "Biostrings")
align("ALVDE", "AVRES", score.matrix = BLOSUM62)
}

