% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{getMotifMatrix}
\alias{getMotifMatrix}
\title{getMotifMatrix}
\usage{
getMotifMatrix(object, simplify = TRUE, filter = FALSE)
}
\arguments{
\item{object}{a MotifSearchResult object.}

\item{simplify}{whether to return an presence/absence matrix.}

\item{filter}{whether remove columns/rows with all zero entries.}
}
\description{
getMotifMatrix
}

