% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{getMotifsBySeq}
\alias{getMotifsBySeq}
\title{getMotifsBySeq}
\usage{
getMotifsBySeq(object)
}
\arguments{
\item{object}{a MotifSearchResult object.}
}
\value{
A list of sequences.
}
\description{
Obtain list with a vector of motifs per sequence.
}

