% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{MotifCompareResult-class}
\alias{MotifCompareResult-class}
\alias{show,MotifCompareResult-method}
\title{MotifCompareResult}
\usage{
\S4method{show}{MotifCompareResult}(object)
}
\description{
MotifCompareResult
}
\section{Slots}{

\describe{
\item{\code{info}}{list with result information.}

\item{\code{alphabet}}{data.frame with alphabet information.}

\item{\code{query}}{data.frame with query motifs information.}

\item{\code{target}}{data.frame with target motifs information.}

\item{\code{probabilities}}{list of motif probabilities.}

\item{\code{matches}}{list of motif scores.}
}}

