% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{MotifSearchResult-class}
\alias{MotifSearchResult-class}
\alias{[,MotifSearchResult-method}
\alias{motifNames}
\alias{motifNames,MotifSearchResult-method}
\alias{nmotif}
\alias{nmotif,MotifSearchResult-method}
\alias{nseq}
\alias{nseq,MotifSearchResult-method}
\alias{pwm}
\alias{pwm,MotifSearchResult-method}
\alias{scores}
\alias{scores,MotifSearchResult-method}
\alias{sequenceNames}
\alias{sequenceNames,MotifSearchResult-method}
\alias{show,MotifSearchResult-method}
\title{MotifSearchResult-class}
\usage{
\S4method{show}{MotifSearchResult}(object)

\S4method{[}{MotifSearchResult}(x, i, j, ..., drop = FALSE)

nmotif(object)

nseq(object)

motifNames(object)

sequenceNames(object)

scores(object)

pwm(object)
}
\arguments{
\item{object}{MotifSearchResult object.}

\item{x}{MotifSearchResult object.}

\item{i}{row number (motif).}

\item{j}{column number (sequence).}

\item{...}{further arguments passed down to method.}

\item{drop}{logical; whether to coerce to vector when number of columns equal one.}
}
\description{
This document described the MotifSearchResult class and methods for storing
results from motif finding methods (e.g. MEME, FIMO and MAST).
}
\section{Slots}{

\describe{
\item{\code{info}}{list with result information.}

\item{\code{sequences}}{AnnotatedDataFrame with sequence information.}

\item{\code{motifs}}{AnnotatedDataFrame with motif information.}

\item{\code{probabilities}}{list of motif probabilities.}

\item{\code{scores}}{list of motif scores.}

\item{\code{ranges}}{RangedData with motif hits.}
}}

