% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motiftools-package.R
\docType{package}
\name{motiftools-package}
\alias{motiftools}
\alias{motiftools-package}
\title{motiftools: Tools to analyze sequence motifs}
\description{
Functions useful to analyze and interpret results from sequence 
    motif tools. Currently focus is on the MEME suite.
}
\author{
\strong{Maintainer}: Diego Diez \email{diego10ruiz@gmail.com}

}
\keyword{internal}
