% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{plotConservationMatrix}
\alias{plotConservationMatrix}
\title{plotConservationMatrix}
\usage{
plotConservationMatrix(x, tree)
}
\arguments{
\item{x}{a matrix as obtained with conservationMatrix()}

\item{tree}{a tree of class phylo (package ape)}
}
\description{
plots a heatmap with the conservation score for each residue
}

