% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{plotConservationMatrix}
\alias{plotConservationMatrix}
\title{plotConservationMatrix}
\usage{
plotConservationMatrix(x, tree, color = "transparent", show.tips = FALSE,
  show.consensus = FALSE, size = 8)
}
\arguments{
\item{x}{a matrix as obtained with conservationMatrix()}

\item{tree}{a tree of class dendrogram or that can be coerced into.}

\item{color}{color for the tile borders (default: transparent).}

\item{show.tips}{logical; whether to show sequence ids.}

\item{show.consensus}{logical; whether to show consensus sequence.}

\item{size}{text size for consensus sequence.}
}
\description{
plots a heatmap with the conservation score for each residue
}

