% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotMatrix}
\alias{plotMatrix}
\title{plotMatrix}
\usage{
plotMatrix(object, tree, col, high, high.col, plot = TRUE)
}
\arguments{
\item{object}{}

\item{tree}{}

\item{col}{}

\item{high}{}

\item{high.col}{}

\item{plot}{}
}
\value{
Returns a grob object invisibly.
}
\description{
plotMatrix
}
\examples{
NULL
}

