% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotMatrix}
\alias{plotMatrix}
\title{plotMatrix}
\usage{
plotMatrix(object, tree, fill, color = "transparent", high, high.col,
  bar.percentage = TRUE, plot = TRUE)
}
\arguments{
\item{object}{}

\item{tree}{}

\item{fill}{}

\item{color}{}

\item{high}{}

\item{high.col}{}

\item{bar.percentage}{}

\item{plot}{}
}
\value{
Returns a grob object invisibly.
}
\description{
plotMatrix
}
\examples{
NULL
}

