% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{plotMotifArchSimilarity}
\alias{plotMotifArchSimilarity}
\title{plotMotifArchSimilarity}
\usage{
plotMotifArchSimilarity(object, raster = FALSE)
}
\arguments{
\item{object}{MotifSearchResult object.}

\item{raster}{logical; whether to use geom_raster() rather that geom_tile().}
}
\description{
Plots heatmap with similarity between sequence motif architectures.
}
