% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomtom.R
\name{readTOMTOM}
\alias{readTOMTOM}
\title{readTOMTOM}
\usage{
readTOMTOM(file, description = NULL)
}
\arguments{
\item{file}{Tomtom file in XML format.}

\item{description}{description for the experiment.}
}
\value{
MotifCompareResult
}
\description{
Read results from tomtom.
}
\examples{
NULL
}
